Fix docker image build error
Why
---
The dockerfile precisely pins the OpenSSL version.  A new version was
released yesterday.  When the new version was released, the old version
became incompatible with the declared Apline version.  The OpenSSL
version only needed to be pinned while a critical vulnerability existed
(from 3.0.0 to 3.0.6).

What
----
Remove OpenSSL version declaration