Prefer stale dialog pod data to invalid results
We have observed get_dialog_pods/0 fail to discover dialog nodes within
the cluster for reasons we don't yet fully understand. In the meantime,
the simplest "fix" is to prefer stale cache data instead of overwrite
it.

The :ignore atom is meant to tell Cachex not to update the cache.

Since this warmer runs every 15 seconds, we think an occasional
intermittent failure will not cause major issues.