Add Clean up non-referred files function
**Problem**

Due to a bug in Reticulum and/or Hubs Client, OwnedFiles that are not
referred from anywhere can be remained as active in the database then the
associated files are not removed. This problem causes a high disk volume
usage pressure.

**Solution**

This commit addresses the problem by adding a function that finds and
deletes such non-referred files. The function might be removed when the
root issue will be resolved.

Non-referred OwnedFile detection algorithm

1. Mark all OwnedFiles as inactive
2. Mark the OwnedFiles as active that are referred by RoomObjects
3. Mark the OwnedFiles as active that are referred by
   AppConfig, Asset, Avatar, AvatarListing, Project, Scene, or SceneListing
4. Delete the inactive OwnedFiles

Note that in this commit we don't delete the non-referred OwnedFiles
and associated files. Instead until we are confident for this operation
we move the non-referred OwnedFile rows to another new table and keep
the associated files. Even if this operation has a bug and wrongly
detects non-referred owned files we can restore. And users will notice
the wrong detection problem by their contents that won't appear
because they are moved from OwnedFile table.

After we test this operation much enough we will update the code
to actually delete the non-referred files in another commit.

**Changes**

- Add a new table that holds non-referred OwnedFile rows
- Add a function that detects non-referred OwnedFiles and
  moves them to the new table
- Add unit tests for them

**TODO in the future**

- Test this operation much enough
- Update the function to actually delete non-referred files