Update Discord API to v10
What: Updates the Discord client to use version 10 of the Discord API and account for its quirks/changes.

Why: The version of the Discord API was very old and will likely stop working in the near future.  Plus it's good to do as general maintenance.

Notes:
Discord seems to waffle between sending us integers and strings for permissions bitfield values depending on what permission level you are (e.g. users with room owner permissions seemed to have integers sent, but non-privileged users seemed to have strings sent), so this had to be accounted for by testing whether the value was a string or not and converting any strings to integers.

Discord has changed and usernames for people are now enforced to be unique and the discriminator numbers aren't appended to the end anymore (Discord bots still use discriminators, but this isn't relevant to us).
https://discord.fandom.com/wiki/Discriminator